Add change avatar functionality to ProfileImage component and related styles
