+++ src/components/userNotifications/UserNotifications.jsx
@@ -15,6 +15,7 @@ import CoinIcon from '../coinIcon/CoinIcon'
 import useCommonStore from '../../hooks/commonStore'
 import uploadFile from "../../services/uploadFile";
 import { NOTIFICATION_TYPES } from '../../consts/messages'
+import { useProfileImage } from '@/hooks/useProfileImage'
 
 const { motion } = require('framer-motion')
 
@@ -22,9 +23,8 @@ moment.locale('es-CO')
 
 const UserNotifications = () => {
   const { userLogged, notifications, setStoreValue } = useCommonStore((state => state))
+  const { handlerInputFile, fileInputRef } = useProfileImage()
   const { uid } = userLogged
-  const [fileUploaded, setFileUploaded] = useState(false);
-  const fileInputRef = useRef(null);
   const [bannerPictureProfile, setBannerPictureProfile] = useState(false);
   // const user = useSelector(state => state.user)
   // const notifications = useSelector(state => state.notifications) //.filter(item => item.closed == 0)
@@ -84,34 +84,6 @@ const UserNotifications = () => {
       handleDeleteNotification(id)
     }
     if (link) Router.push(link)
-  }
-
-  const ToastContent = (message, type) => {
-    return <>
-      <img src="/images/ia/5.png" />
-      <span className={type} onClick={() => setStoreValue('leftMenuBar', { isShow: true })}>{message}</span>
-    </>
-  }
-
-  const handlerInputFile = async (event) => {
-    const value = event.target.files[0]
-    if (value) {
-      setStoreValue('isFullLoading', true)
-      const urlImage = await uploadFile("profile", value, `${uid}`);
-      updateProfileSrv(null, uid, { picture: urlImage })
-        .then(resp => {
-          const { data } = resp
-          setStoreValue('userLogged', { ...userLogged, picture: urlImage })
-          setStoreValue('isFullLoading', false)
-          // if (data.messageTop) setStoreValue('messageTop', data.messageTop)
-          toast(ToastContent(data.messageTop, 'success'))
-          setStoreValue('leftMenuBar', { isShow: false })
-          getNotifications() // Actualizar notificaciones
-        })
-        .catch((err) => {
-          setStoreValue('isFullLoading', false)
-        })
-    }
   }
 
   const container = {
